add: DataHandler class
