+++ covsirphy/analysis/scenario.py
@@ -2,19 +2,17 @@
 # -*- coding: utf-8 -*-
 
 import copy
+from covsirphy.analysis.data_handler import DataHandler
 import numpy as np
 import pandas as pd
 from covsirphy.util.error import deprecate
 from covsirphy.util.plotting import line_plot, box_plot
-from covsirphy.cleaning.term import Term
-from covsirphy.cleaning.jhu_data import JHUData
-from covsirphy.cleaning.population import PopulationData
 from covsirphy.phase.phase_unit import PhaseUnit
 from covsirphy.phase.phase_series import PhaseSeries
 from covsirphy.analysis.param_tracker import ParamTracker
 
 
-class Scenario(Term):
+class Scenario(DataHandler):
     """
     Scenario analysis.
 
@@ -28,26 +26,9 @@ class Scenario(Term):
     """
 
     def __init__(self, jhu_data, population_data, country, province=None, tau=None, auto_complement=True):
-        # Population
-        population_data = self.ensure_instance(
-            population_data, PopulationData, name="population_data")
-        self.population = population_data.value(country, province=province)
-        # Records
-        self.jhu_data = self.ensure_instance(
-            jhu_data, JHUData, name="jhu_data")
-        # Area name
-        self.country = country
-        self.province = province or self.UNKNOWN
-        self.area = JHUData.area_name(country, province)
-        # tau value must be shared
+        super().__init__(
+            jhu_data, population_data, country, province=province, auto_complement=auto_complement)
         self.tau = self.ensure_tau(tau)
-        # Whether complement the number of cases or not
-        self._auto_complement = bool(auto_complement)
-        self._complemented = False
-        # Create {scenario_name: PhaseSeries} and set records
-        self.record_df = pd.DataFrame()
-        self._first_date = None
-        self._last_date = None
         self._init_phase_series()
 
     def __getitem__(self, key):
@@ -58,64 +39,6 @@ class Scenario(Term):
     def __setitem__(self, key, value):
         self.ensure_instance(value, PhaseSeries, name="value")
         self._series_dict[key] = value
-
-    def _init_phase_series(self):
-        """
-        Initialize dictionary of phase series and set records.
-        Only when auto-complement mode, complement records if necessary.
-        """
-        # Set records (complement records, if necessary)
-        self.record_df, self._complemented = self.jhu_data.records(
-            country=self.country, province=self.province,
-            start_date=self._first_date, end_date=self._last_date,
-            population=self.population,
-            auto_complement=self._auto_complement
-        )
-        # First/last date of the records
-        if self._first_date is None:
-            series = self.record_df.loc[:, self.DATE]
-            self._first_date = series.min().strftime(self.DATE_FORMAT)
-            self._last_date = series.max().strftime(self.DATE_FORMAT)
-        # Set main scenario
-        self._series_dict = {
-            self.MAIN: PhaseSeries(
-                self._first_date, self._last_date, self.population
-            )
-        }
-
-    def complement(self, interval=2, max_ignored=100):
-        """
-        Complement the number of recovered cases, if necessary.
-
-        Args:
-            interval (int): expected update interval of the number of recovered cases [days]
-            max_ignored (int): Max number of recovered cases to be ignored [cases]
-
-        Returns:
-            covsirphy.Scenario: self
-        """
-        self.record_df, self._complemented = self.jhu_data.records(
-            country=self.country, province=self.province,
-            start_date=self._first_date, end_date=self._last_date,
-            population=self.population,
-            auto_complement=True, interval=interval, max_ignored=max_ignored
-        )
-        return self
-
-    def complement_reverse(self):
-        """
-        Restore the raw records. Reverse method of covsirphy.Scenario.complement().
-
-        Returns:
-            covsirphy.Scenario: self
-        """
-        self.record_df, self._complemented = self.jhu_data.records(
-            country=self.country, province=self.province,
-            start_date=self._first_date, end_date=self._last_date,
-            population=self.population,
-            auto_complement=False
-        )
-        return self
 
     @property
     def first_date(self):
@@ -145,76 +68,17 @@ class Scenario(Term):
         self._last_date = date
         self._init_phase_series()
 
-    def records(self, show_figure=True, filename=None):
+    def _init_phase_series(self):
         """
-        Return the records as a dataframe.
-
-        Args:
-            show_figure (bool): if True, show the records as a line-plot.
-            filename (str): filename of the figure, or None (show figure)
-
-        Returns:
-            pandas.DataFrame
-                Index:
-                    reset index
-                Columns:
-                    - Date (pd.TimeStamp): Observation date
-                    - Confirmed (int): the number of confirmed cases
-                    - Infected (int): the number of currently infected cases
-                    - Fatal (int): the number of fatal cases
-                    - Recovered (int): the number of recovered cases (> 0)
-
-        Notes:
-            Records with Recovered > 0 will be selected.
-            If complement was performed by Scenario.complement() or Scenario(auto_complement=True),
-            "(complemented)" will be added to the title of the figure.
+        Initialize dictionary of phase series and set records.
+        Only when auto-complement mode, complement records if necessary.
         """
-        df = self.record_df.drop(self.S, axis=1)
-        if not show_figure:
-            return df
-        title = f"{self.area}: Cases over time{' (complemented)' if self._complemented else ''}"
-        line_plot(
-            df.set_index(self.DATE).drop(self.C, axis=1),
-            title,
-            y_integer=True,
-            filename=filename
-        )
-        return df
-
-    def records_diff(self, variables=None, window=7, show_figure=True, filename=None):
-        """
-        Return the number of daily new cases (the first discreate difference of records).
-
-        Args:
-            variables (str or None): variables to show
-            window (int): window of moving average, >= 1
-            show_figure (bool): if True, show the records as a line-plot.
-            filename (str): filename of the figure, or None (show figure)
-
-        Returns:
-            pandas.DataFrame
-                Index:
-                    - Date (pd.TimeStamp): Observation date
-                Columns:
-                    - Confirmed (int): daily new cases of Confirmed, if calculated
-                    - Infected (int):  daily new cases of Infected, if calculated
-                    - Fatal (int):  daily new cases of Fatal, if calculated
-                    - Recovered (int):  daily new cases of Recovered, if calculated
-
-        Notes:
-            @variables can be selected from Confirmed, Infected, Fatal and Recovered.
-            If None was set to @variables, ["Confirmed", "Fatal", "Recovered"] will be used.
-        """
-        variables = self.ensure_list(
-            variables or [self.C, self.F, self.R], candidates=self.VALUE_COLUMNS, name="variables")
-        window = self.ensure_natural_int(window, name="window")
-        df = self.record_df.set_index(self.DATE)[variables]
-        df = df.diff().dropna().rolling(window=window).mean().dropna()
-        if not show_figure:
-            return df
-        title = f"{self.area}: Daily new cases{' (complemented)' if self._complemented else ''}"
-        line_plot(df, title, y_integer=True, filename=filename)
-        return df
+        super()._init_records()
+        self._series_dict = {
+            self.MAIN: PhaseSeries(
+                self._first_date, self._last_date, self.population
+            )
+        }
 
     def _create_tracker(self, name):
         """
